Use YetAnotherConfigLib instead of AutoConfig
Also includes ConfigHandler.java, which turns the config from a JSON file into a TOML file. It also allows for comments, which are defined in the Config class using ConfigComment annotations.

ConfigHandler may be split into its own library after I decouple it from Config, as it would be useful to use in other mods.